feat: implement user login functionality with JWT token generation

+++ Backend/dto/user.go
@@ -16,7 +16,7 @@ type LoginRequest struct {
 }
 
 type UserLoginResponse struct {
-	ID    int64  `json:"id"`
+	ID    uint   `json:"id"`
 	Email string `json:"email"`
 	Token string `json:"token"`
 }
